refactor: auto-split test_security_fixes.py
